requirements.txt
# Claude API client
anthropic>=0.39.0

# GitHub API access
PyGithub>=2.1.1

# Environment variable management
python-dotenv>=1.0.0

# HTTP requests (backup for GitHub API)
requests>=2.31.0
